# docs/swagger.yaml
definitions:
...
  model.CreateSyllabus:
    properties:
      checksum:
        type: string
      contentType:
        type: string
      courseId:
        type: string
      fileName:
        type: string
      fileSize:
        type: integer
      semester:
        type: string
      year:
        type: integer
    type: object
...
  model.Syllabus:
    properties:
      contentType:
        type: string
      courseId:
        type: string
      dateAdded:
        type: integer
      fileName:
        type: string
      fileSize:
        type: integer
      id:
        type: string
      received:
        type: boolean
      semester:
        type: string
      userId:
        type: string
      year:
        type: integer
    type: object
  model.SyllabusLike:
    properties:
      dateReacted:
        type: integer
      dislike:
        type: boolean
      syllabusId:
        type: string
      userId:
        type: string
    type: object
  model.SyllabusReaction:
    properties:
      action:
        type: string
    type: object
  model.UpdateSyllabus:
    properties:
      semester:
        type: string
      year:
        type: integer
    type: object
...
info:
  contact:
    name: Jackie Li
  description: Syllabye API server.
  title: Syllabye API
  version: "1.0"
paths:
...
  /syllabi:
    get:
      parameters:
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Filter by course ID
        in: query
        name: courseId
        type: string
      - description: Filter by year
        in: query
        name: year
        type: integer
      - description: Filter by semester
        in: query
        name: semester
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: size
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Syllabus'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: List syllabi
      tags:
      - Syllabus
    post:
      consumes:
      - application/json
      parameters:
      - description: Syllabus data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateSyllabus'
      responses:
        "201":
          description: Created
          headers:
            Location:
              description: URL to access the created syllabus
              type: string
            X-Presigned-Url:
              description: Presigned URL to upload the syllabus file
              type: string
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: Create a syllabus
      tags:
      - Syllabus
  /syllabi/{syllabusId}:
    delete:
      parameters:
      - description: Syllabus ID
        in: path
        name: syllabusId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: Delete a syllabus
      tags:
      - Syllabus
    get:
      parameters:
      - description: Syllabus ID
        in: path
        name: syllabusId
        required: true
        type: string
      responses:
        "200":
          description: OK
          headers:
            X-Presigned-Url:
              description: Presigned URL to access the syllabus file
              type: string
          schema:
            $ref: '#/definitions/model.Syllabus'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: Get a syllabus
      tags:
      - Syllabus
    patch:
      parameters:
      - description: Syllabus ID
        in: path
        name: syllabusId
        required: true
        type: string
      - description: Updated syllabus data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSyllabus'
      responses:
        "204":
          description: No Content
          headers:
            Location:
              description: URL to access the updated syllabus
              type: string
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: Update a syllabus
      tags:
      - Syllabus
  /syllabi/{syllabusId}/reaction:
    delete:
      parameters:
      - description: Syllabus ID
        in: path
        name: syllabusId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: Remove syllabus reaction
      tags:
      - Syllabus
    post:
      parameters:
      - description: Syllabus ID
        in: path
        name: syllabusId
        required: true
        type: string
      - description: Reaction action
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SyllabusReaction'
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: React to a syllabus
      tags:
      - Syllabus
  /syllabi/{syllabusId}/reactions:
    get:
      parameters:
      - description: Syllabus ID
        in: path
        name: syllabusId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SyllabusLike'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: List syllabus reactions
      tags:
      - Syllabus
  /syllabi/{syllabusId}/sync:
    get:
      parameters:
      - description: Syllabus ID
        in: path
        name: syllabusId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          headers:
            Location:
              description: URL to access the synced syllabus
              type: string
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AWS: []
      summary: Sync a syllabus
      tags:
      - Syllabus
...
securityDefinitions:
  AWS:
    description: 'Provide a valid Bearer token. Example: "Bearer {jwt-token}"'
    in: header
    name: Authorization
    type: apiKey
  Session:
    in: cookie
    name: syllabye.session
    type: apiKey
swagger: "2.0"